Rename 'admin.js' to 'admins.js'
