+++ lib/revisions/routes.js
@@ -6,7 +6,7 @@ const {ensureCodeCommune, ensureCommuneActuelle} = require('../util/middlewares'
 const w = require('../util/w')
 const rawBodyParser = require('../util/raw-body-parser')
 const {createAuthClient} = require('../clients')
-const {createAuthAdmin} = require('../admin')
+const {createAuthAdmin} = require('../admins')
 const {validateBAL} = require('./validate-bal')
 const {fetchRevision, getCurrentRevision, getRevisionsByCommune, createRevision, setFile, getFiles, getFileData, removeFiles, publishRevision, computeRevision, getCurrentRevisions, getRelatedHabilitation} = require('./model')
